Addded StockAccountService

+++ code/FinanceManager.Components/Services/FinancalAccountService.cs
@@ -1,6 +1,7 @@
 ﻿using FinanceManager.Application.Commands.Account;
 using FinanceManager.Domain.Entities.Accounts;
 using FinanceManager.Domain.Entities.Accounts.Entries;
+using FinanceManager.Domain.ValueObjects;
 using Microsoft.Extensions.Logging;
 
 namespace FinanceManager.Components.Services;
@@ -8,25 +9,36 @@ namespace FinanceManager.Components.Services;
 public class FinancalAccountService : IFinancalAccountService
 {
     private readonly BankAccountService _bankAccountService;
+    private readonly StockAccountService _stockAccountService;
     private readonly ILogger<FinancalAccountService> logger;
 
-    public FinancalAccountService(BankAccountService bankAccountService, ILogger<FinancalAccountService> logger)
+    public FinancalAccountService(BankAccountService bankAccountService, StockAccountService stockAccountService, ILogger<FinancalAccountService> logger)
     {
         _bankAccountService = bankAccountService;
+        _stockAccountService = stockAccountService;
         this.logger = logger;
     }
 
     public async Task<bool> AccountExists(int id)
     {
-        var accounts = await _bankAccountService.GetAvailableAccountsAsync();
+        if ((await _bankAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == id)) return true;
+        if ((await _stockAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == id)) return true;
 
-        return accounts.Any(x => x.AccountId == id);
+        return false;
     }
-
-    public async Task AddAccount<T>(T account) where T : BasicAccountInformation
+    public async Task<bool> AccountExists<T>(int id)
     {
         if (typeof(T) == typeof(BankAccount))
-            await _bankAccountService.AddAccountAsync(new AddAccount(account.Name));
+            return (await _bankAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == id);
+        if (typeof(T) == typeof(StockAccount))
+            return (await _stockAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == id);
+
+        return false;
+    }
+    public async Task AddAccount<T>(T account) where T : BasicAccountInformation
+    {
+        if (account is BankAccount) await _bankAccountService.AddAccountAsync(new AddAccount(account.Name));
+        if (account is StockAccount) await _stockAccountService.AddAccountAsync(new AddAccount(account.Name));
     }
 
     public async Task AddAccount<AccountType, EntryType>(string accountName, List<EntryType> data)
@@ -44,18 +56,40 @@ public class FinancalAccountService : IFinancalAccountService
                 }
             }
         }
+        else if (typeof(AccountType) == typeof(StockAccount))
+        {
+            var bankAccountId = await _stockAccountService.AddAccountAsync(new AddAccount(accountName));
+            foreach (var item in data)
+            {
+                if (item is StockAccountEntry stockEntry)
+                {
+                    await _stockAccountService.AddEntryAsync(new AddStockAccountEntry(stockEntry));
+                }
+            }
+        }
     }
 
     public async Task AddEntry<T>(T accountEntry) where T : FinancialEntryBase
     {
-        if (accountEntry is BankAccountEntry bankAccountEntry)
-            await _bankAccountService.AddEntryAsync(new AddBankAccountEntry(bankAccountEntry));
+        switch (accountEntry)
+        {
+            case BankAccountEntry bankAccountEntry:
+                await _bankAccountService.AddEntryAsync(new AddBankAccountEntry(bankAccountEntry));
+                break;
+
+            case StockAccountEntry stockAccountEntry:
+                await _stockAccountService.AddEntryAsync(new AddStockAccountEntry(stockAccountEntry));
+                break;
+
+        }
     }
 
     public async Task<T?> GetAccount<T>(int userId, int id, DateTime dateStart, DateTime dateEnd) where T : BasicAccountInformation
     {
         if (typeof(T) == typeof(BankAccount))
             return await _bankAccountService.GetAccountWithEntriesAsync(id, dateStart, dateEnd) as T;
+        else if (typeof(T) == typeof(StockAccount))
+            return await _stockAccountService.GetAccountWithEntriesAsync(id, dateStart, dateEnd) as T;
 
         return null;
     }
@@ -63,7 +97,13 @@ public class FinancalAccountService : IFinancalAccountService
     public async Task<IEnumerable<T>> GetAccounts<T>(int userId, DateTime dateStart, DateTime dateEnd) where T : BasicAccountInformation
     {
         List<T> result = [];
-        var accounts = await _bankAccountService.GetAvailableAccountsAsync();
+        IEnumerable<AvailableAccount> accounts = [];
+
+        if (typeof(T) == typeof(BankAccount))
+            accounts = await _bankAccountService.GetAvailableAccountsAsync();
+        else if (typeof(T) == typeof(StockAccount))
+            accounts = await _stockAccountService.GetAvailableAccountsAsync();
+
         foreach (var account in accounts)
         {
             T? nextAccount = await GetAccount<T>(userId, account.AccountId, dateStart, dateEnd);
@@ -76,26 +116,45 @@ public class FinancalAccountService : IFinancalAccountService
 
     public async Task<Dictionary<int, Type>> GetAvailableAccounts()
     {
-        var accounts = await _bankAccountService.GetAvailableAccountsAsync();
-        return accounts.ToDictionary(x => x.AccountId, x => typeof(BankAccount));
+        Dictionary<int, Type> result = [];
+
+        foreach (var account in await _bankAccountService.GetAvailableAccountsAsync())
+            result.Add(account.AccountId, typeof(BankAccount));
+
+        foreach (var account in await _stockAccountService.GetAvailableAccountsAsync())
+            result.Add(account.AccountId, typeof(StockAccount));
+
+        return result;
     }
 
     public async Task<DateTime?> GetEndDate(int accountId)
     {
-        return await _bankAccountService.GetYoungestEntryDate(accountId);
+        if ((await _bankAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == accountId))
+            return await _bankAccountService.GetYoungestEntryDate(accountId);
+
+        if ((await _stockAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == accountId))
+            return await _stockAccountService.GetYoungestEntryDate(accountId);
+
+        return null;
     }
     public async Task<DateTime?> GetStartDate(int accountId)
     {
-        return await _bankAccountService.GetOldestEntryDate(accountId);
+        if ((await _bankAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == accountId))
+            return await _bankAccountService.GetOldestEntryDate(accountId);
+
+        if ((await _stockAccountService.GetAvailableAccountsAsync()).Any(x => x.AccountId == accountId))
+            return await _stockAccountService.GetOldestEntryDate(accountId);
+        return null;
     }
 
     public async Task<int?> GetLastAccountId()
     {
-        var bankAccounts = await _bankAccountService.GetAvailableAccountsAsync();
+        List<AvailableAccount> accounts = (await _bankAccountService.GetAvailableAccountsAsync()).ToList();
+        accounts.AddRange((await _stockAccountService.GetAvailableAccountsAsync()).ToList());
 
-        if (!bankAccounts.Any()) return 0;
+        if (!accounts.Any()) return 0;
 
-        return bankAccounts.Max(x => x.AccountId);
+        return accounts.Max(x => x.AccountId);
     }
 
 
@@ -107,22 +166,36 @@ public class FinancalAccountService : IFinancalAccountService
 
     public async Task RemoveAccount(int id)
     {
-        await _bankAccountService.DeleteAccountAsync(new DeleteAccount(id));
+        if (await AccountExists<BankAccount>(id))
+            await _bankAccountService.DeleteAccountAsync(new DeleteAccount(id));
+
+        if (await AccountExists<StockAccount>(id))
+            await _stockAccountService.DeleteAccountAsync(new DeleteAccount(id));
     }
 
-    public async Task RemoveEntry(int accountEntryId, int entryId)
+    public async Task RemoveEntry(int entryId, int accountId)
     {
-        await _bankAccountService.DeleteEntryAsync(accountEntryId, entryId);
+        if (await AccountExists<BankAccount>(accountId))
+            await _bankAccountService.DeleteEntryAsync(accountId, entryId);
+
+        if (await AccountExists<StockAccount>(accountId))
+            await _stockAccountService.DeleteEntryAsync(accountId, entryId);
     }
 
     public async Task UpdateAccount<T>(T account) where T : BasicAccountInformation
     {
-        await _bankAccountService.UpdateAccountAsync(new UpdateAccount(account.AccountId, account.Name));
+        if (account is BankAccount)
+            await _bankAccountService.UpdateAccountAsync(new UpdateAccount(account.AccountId, account.Name));
+
+        if (account is StockAccount)
+            await _stockAccountService.UpdateAccountAsync(new UpdateAccount(account.AccountId, account.Name));
     }
 
     public async Task UpdateEntry<T>(T accountEntry) where T : FinancialEntryBase
     {
         if (accountEntry is BankAccountEntry bankAccountEntry)
             await _bankAccountService.UpdateEntryAsync(bankAccountEntry);
+        else if (accountEntry is StockAccountEntry stockAccountEntry)
+            await _stockAccountService.UpdateEntryAsync(stockAccountEntry);
     }
 }
